finished vending_factory tests
