+++ contracts/splits/src/lib.rs
@@ -3,5 +3,4 @@ mod error;
 pub mod helpers;
 pub mod msg;
 pub mod state;
-
 pub use crate::error::ContractError;
