+++ test-suite/src/vending_factory.rs
@@ -1,0 +1,1 @@
+mod tests;
